Fixing

+++ src/main/java/org/springframework/gresur/web/ContratoController.java
@@ -6,12 +6,18 @@ import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
+import org.springframework.gresur.configuration.services.UserDetailsImpl;
 import org.springframework.gresur.model.Contrato;
+import org.springframework.gresur.model.EncargadoDeAlmacen;
 import org.springframework.gresur.model.Personal;
+import org.springframework.gresur.model.Transportista;
+import org.springframework.gresur.repository.UserRepository;
 import org.springframework.gresur.service.AdministradorService;
 import org.springframework.gresur.service.ContratoService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -34,6 +40,9 @@ public class ContratoController {
 	private AdministradorService admService;
 	
 	@Autowired
+	UserRepository userRepository;
+	
+	@Autowired
 	public ContratoController(ContratoService contratoService) {
 		this.contratoService = contratoService;
 	}
@@ -41,7 +50,6 @@ public class ContratoController {
 	@GetMapping("/{rol}")
 	@PreAuthorize("hasRole('ADMIN')")
 	public ResponseEntity<?> findPersonal(@PathVariable("rol") String rol) {
-		
 		if(rol.equals("TODOS")) {
 			return ResponseEntity.ok(contratoService.findLastContratoAllEmpleados());
 		} else if(rol.equals("ENCARGADO") || rol.equals("DEPENDIENTE") || rol.equals("ADMINISTRADOR") || rol.equals("TRANSPORTISTA")) {
@@ -75,6 +83,9 @@ public class ContratoController {
 			
 			if(result.hasErrors()) {
 				List<FieldError> le = result.getFieldErrors();
+				if(contratoService.findByPersonalNIF(nif) == null) {
+					admService.deleteByNIF(nif);
+				}
 				return ResponseEntity.badRequest().body(le.get(0).getDefaultMessage() + (le.size()>1? " (Total de errores: " + le.size() + ")" : ""));
 			}
 			
@@ -83,6 +94,11 @@ public class ContratoController {
 					Contrato contrato = contratoService.save(cdef);
 					return ResponseEntity.ok(contrato);
 				}catch(Exception e) {
+					System.out.println("HOLAP1");
+					if(contratoService.findByPersonalNIF(nif) == null) {
+						System.out.println("HOLAP2");
+						admService.deleteByNIFPersonal(nif);
+					}
 					return ResponseEntity.badRequest().body(e.getMessage());
 				}	
 			}
@@ -136,17 +152,24 @@ public class ContratoController {
 		
 		String nifFormato = "^[0-9]{8}([A-Z]{1})?";
 		
+		Authentication user = SecurityContextHolder.getContext().getAuthentication();
+		UserDetailsImpl userDetails = (UserDetailsImpl) user.getPrincipal();
+		
 		if(!nif.matches(nifFormato)) {
 			return ResponseEntity.badRequest().body("Formato del NIF invalido");
 		}
-		
 		if(admService.findByNIFPersonal(nif)==null) {
 			return ResponseEntity.badRequest().body("No existe ningun empleado con dicho NIF");
 		}
-		
 		else {
 			try {
-				contratoService.deleteByPersonalNIF(nif);
+				Personal p = userRepository.findByUsername(userDetails.getUsername()).orElse(null).getPersonal();
+				
+				if(p.getNIF().equals(nif)) {
+					return ResponseEntity.badRequest().body("No se puede eliminar a si mismo");
+				}
+				
+				contratoService.deleteByPersonalNIF(nif);	
 				return ResponseEntity.ok("Borrado correctamente");
 			}catch(Exception e) {
 				return ResponseEntity.badRequest().body(e.getMessage());
